+++ hooks/useNotiUserQuery.ts
@@ -1,5 +1,5 @@
 import { useMutation } from "react-query";
-import { ClearNotification } from "./userRequest";
+import { ClearNotification } from "./userNotiRequest";
 
 export function useClearNotification(id: string, alertId: string, jwt: string) {
   return useMutation(() => ClearNotification(id, alertId, jwt));
